requirements.txt
asttokens==2.0.4
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
Cython==0.29.21
decorator==4.4.2
elasticsearch==7.12.0
executing==0.5.4
gensim==4.0.1
graphviz==0.16
icecream==2.1.0
idna==2.10
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
networkx==2.5.1
nltk==3.6.1
numpy==1.20.2
packaging==20.9
pathy==0.4.0
Pillow==8.2.0
preshed==3.0.5
pydantic==1.7.3
Pygments==2.8.1
pyparsing==2.4.7
pytesseract==0.3.7
pytextrank==3.1.1
rake-nltk==1.0.4
regex==2021.4.4
requests==2.25.1
scipy==1.6.2
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.2
srsly==2.4.1
thinc==8.0.2
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
